dom/media/tests/mochitest/test_zmedia_cleanup.html

<!DOCTYPE HTML>
<html>
  <head>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
    <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript" src="pc.js"></script>
  </head>
<body>
<pre id="test">
<script type="application/javascript">
SimpleTest.waitForExplicitFinish();

if ("nsINetworkInterfaceListService" in SpecialPowers.Ci) {
  var utils = getNetworkUtils();
  utils.tearDownNetwork(function() {
    ok(true, 'Successfully teared down network interface');
    SimpleTest.finish();
  }, function() {
    ok(true, 'Network interface was in down state already');
    SimpleTest.finish();
  });
} else {
  ok(true, 'No need to cleanup network interface');
  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
